requirements.txt
streamlit
opencv-python
insightface==0.7.3
numpy
pandas
scipypip